backend/requirements.txt
# FinanceApp Backend Dependencies
# Python 3.10+

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9  # PostgreSQL adapter

# Validation
pydantic[email]>=2.5.0

# Authentication
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0

# Date handling
python-dateutil>=2.8.2

# Development
python-dotenv>=1.0.0
